Made AppNavbar match Figma
